requirements.txt
adal==1.2.2
pandas==0.25.3
matplotlib==3.1.1
pytest-cov==2.8.1
